[REFACTOR] 리뷰에 따라 서비스함수 로직 변경
- member 값에 따른 분기를 서비스 함수에서 수행하도록 변경
- 서비스 함수 변경에 따른 테스트 코드 변경
